DeleteNonPresentLines for MasterDetailRelation

+++ Serenity.Services/RequestHandlers/IntegratedFeatures/MasterDetail/MasterDetailRelationBehavior.cs
@@ -258,7 +258,7 @@ namespace Serenity.Services
 
             var rowIdField = (Field)((row as IIdRow).IdField);
 
-            if (newList.Count == 0)
+            if (newList.Count == 0 && attr.DeleteNonPresentLines)
             {
                 foreach (Row entity in oldList)
                     DeleteDetail(uow, rowIdField.AsObject(entity));
@@ -279,12 +279,15 @@ namespace Serenity.Services
                     newById[idStr] = item;
             }
 
-            foreach (Row item in oldList)
+            if (attr.DeleteNonPresentLines)
             {
-                var id = rowIdField.AsObject(item);
-                var idStr = AsString(id);
-                if (!newById.ContainsKey(idStr))
-                    DeleteDetail(uow, id);
+                foreach (Row item in oldList)
+                {
+                    var id = rowIdField.AsObject(item);
+                    var idStr = AsString(id);
+                    if (!newById.ContainsKey(idStr))
+                        DeleteDetail(uow, id);
+                }
             }
 
             foreach (Row item in newList)
